319: Refactor CLI commands to generate files from definitions
Implemented sorting of source code definitions generation order and
refactoring code to suppy an abstract CLI command to keep repeatable
code at one place
